Start to vendor packages and ignore config/hub
The plan here is to make sure that we always install the same version of
nvim and tmux while I'm making use of the latest features. On MacOS, it
almost works. I can use `--HEAD` to install both packages with homebrew,
but then some things break with tmux since it identifies its version as
`tmux next-3.1`, which is not helpful for powerline. The next step will
be building and installing these to `~/.local`.